osm/gmaps/requirements.txt
requests
matplotlib
python-dotenv
scikit-image